refactor: move loadOptionsFromJson to lib
